% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{makegraph}
\alias{makegraph}
\title{Construct graph}
\usage{
makegraph(df, directed = TRUE, coords = NULL)
}
\arguments{
\item{df}{A data.frame or matrix containing 3 columns: from, to, cost. See details.}

\item{directed}{logical. If FALSE, then all edges are duplicated by inverting 'from' and 'to' nodes.}

\item{coords}{A data.frame or matrix containing all nodes coordinates.}
}
\value{
List
}
\description{
Construct graph
}
\details{
'from' and 'to' are character or numeric vector containing nodes IDs. 
'cost' is a numeric vector describing the cost (e.g time, distance) between each 'from' and 'to' nodes.
coords should not be angles (e.g latitude and longitude), but expressed in a projection system.
}
