% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distance_pair.R
\name{get_distance_pair}
\alias{get_distance_pair}
\title{Compute shortest distance between origin and destination nodes.}
\usage{
get_distance_pair(Graph, from, to, algorithm = "Dijkstra",
  constant = 1, allcores = FALSE)
}
\arguments{
\item{Graph}{An object generated by cppRouting::makegraph() function.}

\item{from}{A vector of one or more vertices from which distances are calculated (origin).}

\item{to}{A vector of one or more vertices (destination).}

\item{algorithm}{character. "Dijkstra" or "A*"}

\item{constant}{numeric. Constant to maintain the heuristic function admissible. 
Default to 1, when cost is expressed in the same unit than coordinates. See details}

\item{allcores}{Logical. If TRUE, all cores are used.}
}
\value{
Vector of shortest distances.
}
\description{
Compute shortest distance between origin and destination nodes.
}
\details{
To perform A*, projected coordinates should be provided in the Graph object.  
To understand how A star algorithm work, see https://en.wikipedia.org/wiki/A*_search_algorithm .
To understand the importance of constant parameter, see the package description : https://github.com/vlarmet/cppRouting
}
\note{
'from' and 'to' must be the same length.
}
