% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_multi_paths.R
\name{get_multi_paths}
\alias{get_multi_paths}
\title{Compute all shortest paths between origin and destination nodes.}
\usage{
get_multi_paths(Graph, from, to)
}
\arguments{
\item{Graph}{An object generated by cppRouting::makegraph() function.}

\item{from}{A vector of one or more vertices from which shortest paths are calculated (origin).}

\item{to}{A vector of one or more vertices (destination).}
}
\value{
List containing lists of shortest paths.
}
\description{
Compute all shortest paths between origin and destination nodes.
}
\details{
get_multi_paths() recursively perform Dijkstra algorithm for each 'from' nodes.
}
\note{
Be aware that if 'from' and 'to' have consequent size, output will require much memory space.
}
